requirements.txt
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
lightgbm==4.3.0 #to train thw RUL prediction model
matplotlib==3.8.2
#seaborn==0.13.2-not used in codebase
reliability==0.8.15
streamlit==1.32.0
plotly==5.19.0
#sqlalchemy-using sqlite3 directly
langchain
langchain-openai
langchain-community
langgraph==0.0.37
chromadb==0.4.24
#apache-airflow- DAG kept for portfolio but not run locally (too heavy)
python-dotenv==1.0.1
#requests==2.31.0 - no http calls in database,so not needed
joblib==1.3.2
sentence-transformers # for huggingface embeddings 
onnxruntime==1.16.3